Add of the draft test folder
Experimentation of an architecture. All the file are in the same folder for now, I will create diferent folder to have a decent strucure soon. You can test the architecture by launching the file TicTacToe.py
I also implement Ben's idea : GuiControler has control all the gui and TicTacToe interact just with the guiControler, and all the gui (grid and menu) interact just with the guiControler.